Move common kubernetes auth test code in a shared module and also run those auth tests with the KubeVirt compute driver.
